+++ src/components/index.js
@@ -1,4 +1,2 @@
-export {default as Header} from "./Header/Header"
 export {default as Container} from "./Container/Container"
-export {default as Footer} from "./Footer/Footer"
-export {default as Hero} from "./Hero/Hero"
+
