+++ src/sectios/Hero/Hero.css
@@ -1,6 +1,6 @@
 
 .hero-main {
-    background: url("../../components/assets/images/banner-bg.jpg") no-repeat top right/cover;
+    background: url("../../assets/images/banner-bg.jpg") no-repeat top right/cover;
     background-color:var(--color-dark);
     color:var(--color-light);
     border-radius: var(--border-radius-primary);
