+++ src/sectios/index.js
@@ -1,0 +1,3 @@
+export {default as Header} from "./Header/Header"
+export {default as Footer} from "./Footer/Footer"
+export {default as Hero} from "./Hero/Hero"
